Move Diary domain codes to Archive domain
Archive 관련 API들은 모두 Archive 도메인으로 변경
1. 달력 조회
2. 일기 리스트 조회
3. 일기 조회 (Diary 도메인에 있는 일기 조회는 Mailbox 도메인 - 우편 조회 시 사용)
