+++ src/main/java/com/umc/btos/src/archive/model/GetDoneRes.java
@@ -1,0 +1,13 @@
+package com.umc.btos.src.archive.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class GetDoneRes {
+    private int doneIdx;
+    private String content;
+}
